requirements.txt
fastapi
uvicorn
pillow
tensorflow
numpy
python-multipart
mediapipe